Crate hybrid_clocks

Implementation of Hybrid Logical Clocks.

This is based on the paper "Logical Physical Clocks and Consistent Snapshots in Globally Distributed Databases". Provides a strictly-monotonic clock that can be used to determine if one event happens-before another.

Structs

Clock
The main clock type.

ManualClock

Timestamp
A value that represents a logical timestamp.

Wall
A clock source that returns wall-clock in nanoseconds.

WallT
Nanoseconds since unix epoch

Enums

Error

Traits

ClockSource
Describes the interface that the inner clock source must provide.